Give every switch a fallback and switch on key numbers as integers
Enable clang-diagnostic-switch, clang-diagnostic-switch-default and
bugprone-switch-missing-default-case together; their 1,239 reports cover
409 switch statements.

Ninety-eight input switches mixed keyboard enumerators with ButtonKey()
gadget IDs, which are not KeyNumType values; they now switch on the integer
key number. The 311 switches without a default get default: break; before
their closing brace, the fallback GCC's -Wswitch-default already requires.
No case body changed. The switch inside GoogleTest's EXPECT_DEATH is
suppressed at its one use.

The full strict build of both games is clean and all 242 tests pass.

+++ src/ra/expand.cc
@@ -459,7 +459,7 @@ bool Expansion_Dialog(bool bCounterstrike)  //	If not bCounterstrike, then this
     }
 
     KeyNumType input = buttons->Input();
-    switch (input) {
+    switch (static_cast<int>(input)) {
       case ButtonKey(200):
         Whom = list.Current_Object()->House;
         Scen.Scenario = list.Current_Object()->Scenario;
